added widget: display weekly weather

+++ lib/models/weather_forecast_daily.dart
@@ -143,7 +143,7 @@ class WeatherList {
         weather.add(Weather.fromJson(v));
       });
     }
-    speed = json['speed'];
+    speed = json['speed'].toDouble();
     deg = json['deg'];
     clouds = json['clouds'];
     //rain = json['rain'];
